--- app/Http/Controllers/AttributeController.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-
-class AttributeController extends Controller
-{
-    //
-}
