--- app/Http/Controllers/RoleController.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-
-class RoleController extends Controller
-{
-    //
-}
