--- app/Http/Controllers/SupplierController.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-
-class SupplierController extends Controller
-{
-    //
-}
